Try resolving ansible_collections "filters" to support namespaced filters.
The solution presented here is based on globbing and not particularly beautiful,
but I didn't see an obvious way to retrieve the filters from Ansible itself; that means
custom filters inside Ansible project directories are likely not loaded correctly either,
but it ought to work for the ones installed with pip.

We should probably do something similar for the "tests" too if this approach works for @pyhedgehog.

fixes #20
see #19 as well.